+++ lizard_esf/models.py
@@ -9,6 +9,8 @@ from django.db import models
 from lizard_area.models import Area
 from treebeard.mp_tree import MP_Node
 from lizard_fewsnorm.models import TimeSeriesCache
+from lizard_fewsnorm.models import Location
+from lizard_fewsnorm.models import GeoLocationCache
 
 from lizard_security.models import DataSet
 
@@ -122,6 +124,21 @@ class Configuration(MP_Node):
         ('L', 'Logical'),
     )
 
+    RELATED_LOCATION_CHOICES = (
+        ('l_esf2', 'l_esf2'),
+        ('l_zicht', 'l_zicht'),
+        ('l_extin', 'l_extin'),
+        ('l_wathte', 'l_wathte'),
+        ('l_kwel', 'l_kwel'),
+        ('l_wegz', 'l_wegz'),
+        ('l_chlor', 'l_chlor'),
+        ('l_oppq', 'l_oppq'),
+        ('l_grq', 'l_grq'),
+        ('l_stovq', 'l_stovq'),
+        ('l_debiet', 'l_debiet'),
+        ('l_gebied', 'l_gebied'),
+        )
+
     name = models.CharField(max_length=128)
     code = models.CharField(max_length=128)
     is_main_esf = models.IntegerField(null=True, blank=True, choices=ESF_CHOICES,
@@ -132,8 +149,8 @@ class Configuration(MP_Node):
                                      help_text='Grafiek groep die aan gezet moet worden bij openen van dit element')
     source_name = models.CharField(max_length=128,
                                    null=True,
-                                    blank=True,
-                                     help_text='Meta data over de Bron waar de automatisch berekende waarde vandaan komt')
+                                   blank=True,
+                                   help_text='Meta data over de Bron waar de automatisch berekende waarde vandaan komt')
     expanded = models.BooleanField(default=False,
                                      help_text='Wordt volgens mij niet meer gebruikt')
 
@@ -151,8 +168,13 @@ class Configuration(MP_Node):
                                     blank=True,
                                     help_text='link naar de basis parameter, die bij expert parameters wordt overschreven')
 
-
     #settings for getting actual values from Fews
+    related_location = models.CharField(
+        max_length=20,
+        choices=RELATED_LOCATION_CHOICES,
+        null=True, blank=True,
+        help_text=('When filled in, it will take the location '
+                   'from this related_location column'))
     default_parameter_code_manual_fews = models.CharField(max_length=256,
                                      blank=True,
                                      default='',
@@ -197,7 +219,7 @@ class Configuration(MP_Node):
         )
 
     def __unicode__(self):
-        return "%s (%i)"% (self.name, self.id)
+        return "%s (%i)" % (self.name, self.id)
 
 
 class AreaConfiguration(models.Model):
@@ -247,6 +269,7 @@ class AreaConfiguration(models.Model):
         options -- type dict, containing query parameters
                    for TimeSeriesCache
         """
+
         parametercache = self.get_manual_fews_code_part(
             self.configuration.default_parameter_code_manual_fews, 0)
         modulecache = self.get_manual_fews_code_part(
@@ -265,17 +288,50 @@ class AreaConfiguration(models.Model):
         if qualifiersetcache is not None and qualifiersetcache != "":
             options["qualifiersetcache__ident"] = qualifiersetcache
 
+    def create_options(self):
+        """Create options dict to query fews database."""
+        options = {}
+        if self.configuration.related_location is None:
+            options.update({"geolocationcache__ident": self.area.ident})
+        else:
+            related_location_id = self.get_related_location(
+                self.area.ident, self.configuration.related_location)
+            if related_location_id is not None:
+                options.update({"geolocationcache__ident": related_location_id})
+            else:
+                logger.error("Related location is not porperly configured for"\
+                                 "esf-configuration '%s'." % self.configuration.name)
+        return options
+
     # Will sometimes give an unicode error.
     # def __unicode__(self):
     #     return '%s %s' % (self.area, self.configuration)
+    def fews_norm_source(self, location):
+        if location and location.fews_norm_source:
+            return location.fews_norm_source
+        else:
+            return None
+
+    def get_related_location(self, ident, related_location_field):
+        """Retrieve id of related location from fews db."""
+        locations = GeoLocationCache.objects.filter(ident=ident)
+        if locations.exists():
+            location = locations[0]
+            source = self.fews_norm_source(location)
+            fews_location = Location.from_raw(
+                schema_prefix=source.database_schema_name,
+                ident=location.ident,
+                related_location=related_location_field).using(
+                source.database_name)[0]
+            return fews_location.related_location
 
     def get_mydump(self, full_config):
         """
         Dump as dict.
         """
-        manual_value = self.manual_value;
+        manual_value = self.manual_value
         if self.configuration.value_type.name == 'text':
-            manual_value = self.manual_text_value;
+            manual_value = self.manual_text_value
         output = {
             'id': self.id,
             'config_id': self.configuration.id,
@@ -297,7 +353,7 @@ class AreaConfiguration(models.Model):
             output['auto_value'] = None
             self.manual = None
         elif self.configuration.configuration_type.name in ['expert_setting', 'info_setting']:
-            ref_rec = full_config.filter(configuration = self.configuration.base_config)
+            ref_rec = full_config.filter(configuration=self.configuration.base_config)
 
             if len(ref_rec) > 0:
                 output['auto_value'] = ref_rec[0].manual_value
@@ -305,7 +361,7 @@ class AreaConfiguration(models.Model):
                 output['auto_value'] = 'ref mist'
         else:
             #get value from fews
-            options = {"geolocationcache__ident": self.area.ident}
+            options = self.create_options()
             self.update_filter_options(options)
             if not 'parametercache__ident' in options:
                 output['auto_value'] = -999
@@ -425,10 +481,10 @@ def get_data_main_esf(area):
             if config_dump['manual'] == 0:
                 rec['value'] = config_dump['auto_value']
                 rec['source'] = 'auto'
-                rec['source_info'] = '' #+ configurationdate?
+                rec['source_info'] = ''  # + configurationdate?
                 # Bad configuration (?) leads to auto_value -999 and no auto_value_ts.
                 rec['date'] = config_dump.get('auto_value_ts', None)
-                rec['stars'] = 0 #auto_status.value
+                rec['stars'] = 0  # auto_status.value
                 rec['stars_comment'] = None
             else:
                 rec['value'] = 0
@@ -448,7 +504,7 @@ def get_data_main_esf(area):
         data[config.configuration.is_main_esf] = rec
 
     output = []
-    for i in range(1,10):
+    for i in range(1, 10):
         if i in data:
             output.append(data[i])
         else: